alx_travel_app/requirement.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
drf-spectacular==0.27.0
django-environ==0.11.2
mysqlclient==2.2.0
python-decouple==3.8
setuptools>=68.0.0
django-filter==23.5
requests==2.31.0
redis==5.0.1
amqp==5.2.0
kombu==5.3.4
vine==5.1.0
billiard==4.2.0